Move shareable code from DeleteBookConfirmationDialog to a base class
We'll need other dialogs to have double confirmation.
